Windows build now working, switch from plog to spdlog

+++ cpp/ccxt/base/functions/time.h
@@ -9,9 +9,9 @@ namespace ccxt
     double now();
 
     // Returns the unix timestamp in yymmmdd format, assuming the timestamp is in milliseconds and UTC timezone.
-    std::string yymmdd(int timestamp, std::string infix = "");
+    std::string yymmdd(uint64_t timestamp, std::string infix = "");
 
     // Returns the unix timestamp in ISO8601 format, assuming the timestamp is in milliseconds and UTC timezone.
-    std::string iso8601(int timestamp);
+    std::string iso8601(uint64_t timestamp);
 
 } // namespace ccxt
